requirements.txt
# 核心依赖
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.46.0
pillow>=9.0.0
numpy>=1.21.0
tqdm>=4.62.0

# 配置文件
pyyaml>=6.0

# 工具
addict>=2.4.0
easydict>=1.9

# 可视化（可选）
matplotlib>=3.5.0
seaborn>=0.11.0

# DeepSeek-OCR 依赖
einops>=0.6.0

# 多卡训练
accelerate>=0.20.0

# DeepSpeed 多卡训练（高级内存优化和分布式训练）
deepspeed>=0.12.0

